Refactoring - Add docstring, latest experiments - Delete unused code - Update dependencies
