+++ corsmal_challenge/utils.py
@@ -1,6 +1,3 @@
-"""
-https://github.com/OtsuKotsu/MLExp/blob/main/mlexp/utils/reproducibility/seed.py
-"""
 import random
 
 import numpy
@@ -14,9 +11,3 @@ def fix_random_seeds(specified_seed: int = SEED):
     numpy.random.seed(seed=specified_seed)
     torch.manual_seed(seed=specified_seed)
     torch.cuda.manual_seed_all(seed=specified_seed)
-
-
-def behave_deterministically():
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-    torch.use_deterministic_algorithms(True)
