+++ env.yml
@@ -49,11 +49,11 @@ dependencies:
     - fonttools==4.28.5
     - importlib-metadata==4.2.0
     - iniconfig==1.1.1
-    - ipykernel==6.6.1
+    - ipykernel==6.7.0
     - ipython==7.31.0
     - isort==5.10.1
     - jedi==0.18.1
-    - jupyter-client==7.1.0
+    - jupyter-client==7.1.1
     - jupyter-core==4.9.1
     - kiwisolver==1.3.2
     - matplotlib==3.5.1
@@ -88,7 +88,7 @@ dependencies:
     - tomli==1.2.3
     - torch==1.8.2+cu111
     - torchaudio==0.8.2
-    - torchinfo==1.6.1
+    - torchinfo==1.6.3
     - torchvision==0.9.2+cu111
     - tornado==6.1
     - traitlets==5.1.1
