+++ pyproject.toml
@@ -8,15 +8,6 @@ authors = ["Your Name <you@example.com>"]
 python = ">=3.7.4,<3.11"
 numpy = "^1.21.5"
 matplotlib = "^3.5.1"
-#Cython = "^0.29.26"
-#opencv-python = "^4.5.5.62"
-#Pillow = "^8.4.0"
-#scikit-image = "^0.19.1"
-#scikit-learn = "^1.0.2"
-#scipy = "^1.7.3"
-#torch = { url = "https://download.pytorch.org/whl/lts/1.8/cu102/torch-1.8.2%2Bcu102-cp37-cp37m-linux_x86_64.whl" }
-#torchvision = { url = "https://download.pytorch.org/whl/lts/1.8/cu102/torchvision-0.9.2%2Bcu102-cp37-cp37m-linux_x86_64.whl"}
-#torchaudio = { url = "https://download.pytorch.org/whl/lts/1.8/torchaudio-0.8.2-cp37-cp37m-linux_x86_64.whl" }
 torch = { url = "https://download.pytorch.org/whl/lts/1.8/cu111/torch-1.8.2%2Bcu111-cp37-cp37m-linux_x86_64.whl" }
 torchvision = { url = "https://download.pytorch.org/whl/lts/1.8/cu111/torchvision-0.9.2%2Bcu111-cp37-cp37m-linux_x86_64.whl"}
 torchaudio = { url = "https://download.pytorch.org/whl/lts/1.8/torchaudio-0.8.2-cp37-cp37m-linux_x86_64.whl" }
